refactor(cmd): split catalog.go into list and run files
